# JS/views/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "version": "3.0",
        "title":"Kosarkaski klubovi",
        "description": "Košarkaški klubovi iz NBA lige s 3 pripadna igrača i trenerom ",
        "contact": {
            "name": "Borna Radojčić",
            "email": "br52543@fer.hr"
        },
        "license": {
            "name": "CC BY-NC-SA 4.0",
            "url": "https://github.com/Borna0210/or-lab/blob/master/LICENSE"
                }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Razvojni server"
        }
    ],
    "paths": {
        "/kosarka": {
            "get": {
                "summary": "Dohvaćanje svih košarkaških klubova s pripadnim igračima i trenerom",
                "description": "Vraća prirodno spojene tablice iz baze (klub, igrac i trener)",
                "responses": {
                    "200": {
                        "description": "Fetched all data from DB",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type":"array",
                                    "$ref": "#/components/schemas/kosarka",
                                    "example":"localhost:3000/kosarka"
                                }
                            }
                        }
                        
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        },
...
        "/add":{
            "post": {
                "summary": "Novi unos u bazu podataka",
                "description": "Ubacuje novi klub, pripadnog igrača i trenera u bazu podataka",
                "requestBody":{
                    "description": "Sve informacije o klubu, igračima i treneru za dodati u bazu",
                    "content": {
                        "application/json":{
                            "schema":{
                                "$ref":"#/components/schemas/kosarka"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "DB value is now added"
                    },
                    "400": {
                        "description": "You haven't listed all values or you are trying to put a player in a squad with the jersey number that another player has"
                    }
                }
            }
        },
...
        "/openapi":{
            "get": {
                "summary": "Dohvaćanje ovog OpenAPI-ja",
                "description": "Vraća tekst ove OpenAPI dokumentacije",
                "responses": {
                    "200": {
                        "description": "Fetched OpenAPI data",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type":"array",
                                    "$ref": "#/components/schemas/openapi",
                                    "example":"localhost:3000/openapi"
                                }
                            }
                        }
                        
                    },
                    "404": {
                        "description": "OpenAPI data not found on device"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "kosarka":{
                "type": "object",
                "required": [
                    "idklub",
                    "imeklub",
                    "arenaklub",
                    "savdrzava",
                    "godosnutka",
                    "imeigrac",
                    "prezimeigrac",
                    "brojdresa",
                    "pozicijaigrac",
                    "uni_hs_club",
                    "imetrener",
                    "prezimetrener"
                ],
                "properties": {
                    "idklub":{
                        "type": "integer",
                        "example":2
                    },
                    "imeklub":{
                        "type": "string",
                        "example":"Golden State Warriors"
                    },
                    "arenaklub":{
                        "type": "string",
                        "example":"Chase Center"
                    },
                    "savdrzava":{
                        "type": "string",
                        "example":"California"
                    },
                    "godosnutka":{
                        "type": "integer",
                        "example":1946
                    },
                    "imeigrac":{
                        "type": "string",
                        "example":"Stephen"
                    },
                    "prezimeigrac":{
                        "type": "string",
                        "example":"Curry"
                    },
                    "brojdresa":{
                        "type": "integer",
                        "example":30
                    },
                    "pozicijaigrac":{
                        "type": "string",
                        "example":"PG"
                    },
                    "uni_hs_club":{
                        "type": "string",
                        "example":"Davidson"
                    },
                    "imetrener":{
                        "type": "string",
                        "example":"Steve"
                    },
                    "prezimetrener":{
                        "type": "string",
                        "example":"Kerr"
                    }
                }
            },
...
            "openapi":{
                "type": "string",
                "example":"Teško je dati primjer ovoga, ali to je ovaj dokument pa ga možete isprobati"
           
            }
        }
    }
}